# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_ArtemS18_ShortURL-API_internal_delivery_restapi_utils.BaseErrorResponse:
    properties:
      error: {}
      success:
        type: boolean
    type: object
  github_com_ArtemS18_ShortURL-API_internal_usecase_dto.CreateSlugRequest:
    properties:
      url:
        type: string
    required:
    - url
    type: object
  github_com_ArtemS18_ShortURL-API_internal_usecase_dto.CreateSlugResponse:
    properties:
      slug:
        type: string
    type: object
  github_com_ArtemS18_ShortURL-API_internal_usecase_dto.GetURLResponse:
    properties:
      url:
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8000
info:
  contact: {}
  description: Created by Artemii in 2026
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: ShortURL API
  version: "1.0"
paths:
  /slugs:
    post:
      description: Возвращает короткую ссылку для заданного URL
      parameters:
      - description: URL для сокращения
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ArtemS18_ShortURL-API_internal_usecase_dto.CreateSlugRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ при нахождении соответстующей короткой ссылки
            в бд
          schema:
            $ref: '#/definitions/github_com_ArtemS18_ShortURL-API_internal_usecase_dto.CreateSlugResponse'
        "201":
          description: Успешный ответ при создании новой короткой ссылки
          schema:
            $ref: '#/definitions/github_com_ArtemS18_ShortURL-API_internal_usecase_dto.CreateSlugResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/github_com_ArtemS18_ShortURL-API_internal_delivery_restapi_utils.BaseErrorResponse'
      summary: Создание короткой ссылки
      tags:
      - slug
  /slugs/{slug}:
    get:
      description: Возвращает оригинальный URL для заданной короткой ссылки
      parameters:
      - description: Короткая ссылка
        in: path
        name: slug
        required: true
        type: string
      - default: true
        description: Если указан true, то будет выполнин редирект на URL соответсвующий
          вводимой короткой ссылки
        in: query
        name: redirect
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ, если выбран redirect=false
          schema:
            $ref: '#/definitions/github_com_ArtemS18_ShortURL-API_internal_usecase_dto.GetURLResponse'
        "404":
          description: URL не найден
          schema:
            $ref: '#/definitions/github_com_ArtemS18_ShortURL-API_internal_delivery_restapi_utils.BaseErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/github_com_ArtemS18_ShortURL-API_internal_delivery_restapi_utils.BaseErrorResponse'
      summary: Получение URL по короткой ссылке
      tags:
      - slug
swagger: "2.0"